+++ pkg/data/domain/nodepool/aws.go
@@ -7,7 +7,7 @@ import (
 	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta1"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )