+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix flaky nodepool, orgs, and clusters tests caused by timing race in AGE column rendering.
+
 ## [4.12.0] - 2026-03-04
 
 ### Added